Added in lighting to the main scene.
This commit includes changes to some of the lighting prefabs,
which for whatever reason were set up incorrectly.
I also added a script that will auto-generate most of the
lighting geometry. This will be useful for later levels we create.

I changed the player collider, but I don't think this was necessary;
A weird light clipping issue was occuring, but I think it's better to
center the torch light on the player anyways, and it doesn't look
too weird. Otherwise, we're probably going to have to choose between
a large hitbox or lighting not on the torch. Other torches hanging on
the walls shouldn't have this issue, since they don't deal with
collision.